+++ SetApplication/SetApplication.vcxproj
@@ -146,6 +146,9 @@
       <Project>{c1b069dd-71c8-42d0-a4c7-5d38c5e01ed0}</Project>
     </ProjectReference>
   </ItemGroup>
+  <ItemGroup>
+    <ClInclude Include="Classes.h" />
+  </ItemGroup>
   <Import Project="$(VCTargetsPath)\Microsoft.Cpp.targets" />
   <ImportGroup Label="ExtensionTargets">
   </ImportGroup>
